+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## unreleased
+## WIP
+### Changed
+- Updated auto namer for new default name conflict resolution
+    - https://github.com/StateSmith/StateSmith-drawio-plugin/issues/32
 
 ---